separating logic and views
