+++ loading.kv
@@ -1,6 +1,17 @@
 #:kivy 2.1.0
+#:import Factory kivy.factory.Factory
 
-<Content>:
+<Loading>:
+    title: ""
+    size_hint_x: None
+    size_hint_y: None
+    width: "250dp"
+    type: "custom"
+    auto_dismiss: False
+    content_cls: Factory.Content()
+
+
+<Content@BoxLayout>:
     orientation: "vertical"
     size_hint_y: None
     height: "90dp"
